Create Python-Code
Make sure to replace 'path/to/your/file.csv' with the actual path to your CSV file. When you run this script, it will automatically read the URLs from the CSV file, generate the sitemap XML structure, and save it as sitemap.xml in the same directory as the script.